Added MappedSuperClass to allow extensions.
