+++ autoload_classmap.php
@@ -3,7 +3,8 @@
 return array(
     'ZfcUserDoctrineMongoODM\Module'                            => __DIR__ . '/Module.php',
     'ZfcUserDoctrineMongoODM\Document\User'                     => __DIR__ . '/src/ZfcUserDoctrineMongoODM/Document/User.php',
+    'ZfcUserDoctrineMongoODM\Document\UserMappedSuperClass'     => __DIR__ . '/src/ZfcUserDoctrineMongoODM/Document/UserMappedSuperClass.php',
     'ZfcUserDoctrineMongoODM\Event\ResolveTargetEntityListener' => __DIR__ . '/src/ZfcUserDoctrineMongoODM/Event/ResolveTargetEntityListener.php',
-    'ZfcUserDoctrineMongoODM\Mapper\User'                       => __DIR__ . '/src/ZfcUserDoctrineMongoODM/Mapper/User.php',
+    'ZfcUserDoctrineMongoODM\Mapper\UserMongoDb'                => __DIR__ . '/src/ZfcUserDoctrineMongoODM/Mapper/UserMongoDb.php',
     'ZfcUserDoctrineMongoODM\Options\ModuleOptions'             => __DIR__ . '/src/ZfcUserDoctrineMongoODM/Options/ModuleOptions.php'
 );
