--- config/xml/ZfcUserDoctrineMongoODM.Document.UserMeta.mongodb.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
-        http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-                  
-    <entity name="ZfcUserDoctrineMongoODM\Entity\UserMeta" table="user_meta">
-
-    </entity>
-
-</doctrine-mapping>
