+++ src/ZfcUserDoctrineMongoODM/Document/User.php
@@ -2,7 +2,5 @@
 
 namespace ZfcUserDoctrineMongoODM\Document;
 
-use ZfcUser\Entity\User as ZfcUserEntity;
-
-class User extends ZfcUserEntity
+class User extends UserMappedSuperClass
 {}
