--- src/ZfcUserDoctrineMongoODM/Document/UserMeta.php
@@ -1,6 +1,0 @@
-<?php
-
-namespace ZfcUserDoctrineMongoODM\Document;
-
-class UserMeta extends UserMetaMappedSuperClass
-{}
