--- src/ZfcUserDoctrineMongoODM/Document/UserMetaMappedSuperClass.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace ZfcUserDoctrineMongoODM\Document;
-
-use ZfcUser\Model\UserMeta;
-
-class UserMetaMappedSuperClass extends UserMeta
-{}
